refactor: move userSelector to auth
